In 'log' snippet, put cursor inside quotes
Previously, triggering the 'log' snippet would generate two
double-quotes as a placeholder parameter to the function, as in
console.log(""). But this meant that the quotes were selected. To type
text inside the quotes, you'd have to deselect the text manually, then
move your cursor into the quotes. This fixes that problem.